+++ mex/extractors/igs/model.py
@@ -3,17 +3,17 @@ from typing import Any
 from mex.common.models import BaseModel
 
 
-class IGSSchemas(BaseModel):
+class IGSSchema(BaseModel):
     """Model class for IGS Schemas."""
 
 
-class IGSEnum(IGSSchemas):
-    """Model class for IGS Enum."""
+class IGSEnumSchema(IGSSchema):
+    """Model class for IGS Enum Schemas."""
 
     enum: list[str]
 
 
-class IGSProperties(IGSSchemas):
-    """Model class for IGS Properties."""
+class IGSPropertiesSchema(IGSSchema):
+    """Model class for IGS Properties Schemas."""
 
     properties: dict[str, Any]